+++ query-jpa/src/main/java/com/fedorizvekov/query/jpa/model/converter/GenderConverter.java
@@ -5,16 +5,16 @@ import javax.persistence.Converter;
 import com.fedorizvekov.query.jpa.model.enums.Gender;
 
 @Converter(autoApply = true)
-public class GenderConverter implements AttributeConverter<Gender, Byte> {
+public class GenderConverter implements AttributeConverter<Gender, Short> {
 
     @Override
-    public Byte convertToDatabaseColumn(Gender gender) {
+    public Short convertToDatabaseColumn(Gender gender) {
         return gender.getValue();
     }
 
 
     @Override
-    public Gender convertToEntityAttribute(Byte genderValue) {
+    public Gender convertToEntityAttribute(Short genderValue) {
         return Gender.fromValue(genderValue);
     }
